requirements.txt
langchain[groq]
langchain-community
langchainhub
langgraph

fastapi[standard]
python-dotenv
ruff

clickhouse-connect
sqlalchemy-clickhouse
